refactor: improving types
