+++ packages/discovery-react-components/src/components/StructuredQuery/utils/index.ts
@@ -1,9 +1,3 @@
 export { getFieldNames } from './getFieldNames';
 export { getNewId } from './getNewId';
 export { stringifyStructuredQuerySelection } from './stringifyStructuredQuerySelection';
-export {
-  Group,
-  StructuredQuerySelection,
-  OperatorDropdownSelectedItem,
-  RuleGroupDropdownSelectedItem
-} from './structuredQueryInterfaces';
